Rename all instances of covsonar(2) to sonar, mainly in the backend. Rename pg database covsonar to sonar if it exists
