+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
     rev: v0.6.2
     hooks:
       - id: zimports
-        args: ["--keep-unused", "-m", "rest_api,covsonar_backend,permission_model"]
+        args: ["--keep-unused", "-m", "rest_api,sonar_backend,permission_model"]
         exclude: ^apps/backend/rest_api/migrations/
   - repo: https://github.com/psf/black
     rev: 25.1.0
