+++ apps/backend/conf/docker/common.env
@@ -20,7 +20,7 @@ SONAR_EXTERNAL=./work/sonar
 LOG_LEVEL=DEBUG
 
 ## postgres
-POSTGRES_DB=covsonar
+POSTGRES_DB=sonar
 POSTGRES_SCHEMA=public
 POSTGRES_USER=postgres
 POSTGRES_HOST=sonar-db
