+++ apps/backend/conf/docker/dev.env
@@ -2,7 +2,7 @@
 DEBUG=TRUE
 ALLOWED_HOSTS=127.0.0.1,0.0.0.0,localhost
 CORS_ALLOWED_ORIGINS=http://localhost:5173,http://localhost:8000
-DJANGO_SETTINGS_MODULE=covsonar_backend.settings
+DJANGO_SETTINGS_MODULE=sonar_backend.settings
 LOG_PATH=/logs
 PROPERTY_BATCH_SIZE=10000
 START_DJANGO_SERVER="python manage.py collectstatic --noinput && python manage.py runserver 0.0.0.0:9080"
