+++ apps/backend/conf/docker/prod.env
@@ -2,10 +2,10 @@
 DEBUG=FALSE
 ALLOWED_HOSTS=localhost,127.0.0.1
 CORS_ALLOWED_ORIGINS=http://localhost:8000
-DJANGO_SETTINGS_MODULE=covsonar_backend.settings
+DJANGO_SETTINGS_MODULE=sonar_backend.settings
 LOG_PATH=/logs
 PROPERTY_BATCH_SIZE=1000
 GUNI_THREADS=2
 GUNI_WORKERS=2
-START_DJANGO_SERVER="python manage.py collectstatic --noinput && gunicorn --workers ${GUNI_WORKERS} --threads ${GUNI_THREADS} --bind 0.0.0.0:9080 covsonar_backend.wsgi:application"
+START_DJANGO_SERVER="python manage.py collectstatic --noinput && gunicorn --workers ${GUNI_WORKERS} --threads ${GUNI_THREADS} --bind 0.0.0.0:9080 sonar_backend.wsgi:application"
 POSTGRES_CONFIG_FILE=prod.conf
