+++ apps/backend/conf/docker/sonar-cli-action.env
@@ -2,7 +2,7 @@ POSTGRES_INIT_DB=./conf/initdb-test/
 DEBUG=TRUE
 ALLOWED_HOSTS=127.0.0.1,0.0.0.0,localhost
 CORS_ALLOWED_ORIGINS=http://localhost:5173
-DJANGO_SETTINGS_MODULE=covsonar_backend.settings
+DJANGO_SETTINGS_MODULE=sonar_backend.settings
 LOG_PATH=/logs
 SAMPLE_BATCH_SIZE=20
 PROPERTY_BATCH_SIZE=10000
