+++ apps/backend/scripts/linux/dev-run-celery-tasks.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-scripts/linux/dc-dev.sh run --rm sonar-backend celery -A covsonar_backend worker --loglevel=info --without-gossip --without-mingle --without-heartbeat -Ofair
+scripts/linux/dc-dev.sh run --rm sonar-backend celery -A sonar_backend worker --loglevel=info --without-gossip --without-mingle --without-heartbeat -Ofair
