+++ apps/backend/scripts/win/dev-run-celery-tasks.ps1
@@ -1,5 +1,5 @@
 try {
-    scritps\win\dc-dev.ps1 run --rm sonar-backend celery -A covsonar_backend worker --loglevel=info --without-gossip --without-mingle --without-heartbeat -Ofair
+    scritps\win\dc-dev.ps1 run --rm sonar-backend celery -A sonar_backend worker --loglevel=info --without-gossip --without-mingle --without-heartbeat -Ofair
 }
 catch {
     Write-Output "Error occurred while running the script - make sure to run the script from the root directory of the project!"
