+++ apps/backend/sonar_backend/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for covsonar_backend project.
+ASGI config for sonar_backend project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "covsonar_backend.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "sonar_backend.settings")
 
 application = get_asgi_application()
