+++ apps/backend/sonar_backend/celery.py
@@ -10,9 +10,9 @@ from celery import Celery
 from django.conf import settings
 
 # Set the default Django settings module for the 'celery' program.
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "covsonar_backend.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "sonar_backend.settings")
 
-app = Celery("covsonar_backend")  # Replace 'your_project' with your project's name.
+app = Celery("sonar_backend")  # Replace 'your_project' with your project's name.
 
 # Configure Celery using settings from Django settings.py.
 app.config_from_object("django.conf:settings", namespace="CELERY")
