+++ apps/backend/sonar_backend/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for covsonar_backend project.
+Django settings for sonar_backend project.
 
 Generated by 'django-admin startproject' using Django 4.2.4.
 
@@ -17,7 +17,7 @@ from pathlib import Path
 
 import environ
 
-from covsonar_backend.utils import CustomisedJSONFormatter
+from sonar_backend.utils import CustomisedJSONFormatter
 
 # Initialise environment variables
 
@@ -44,7 +44,7 @@ env = environ.Env(
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 PROJECT_ROOT = Path(__file__).resolve().parent
-STATIC_ROOT = BASE_DIR / "covsonar2_server/test/static"
+STATIC_ROOT = BASE_DIR / "sonar_server/test/static"
 
 # Take environment variables from .env file
 environ.Env.read_env(BASE_DIR / ".env")
@@ -93,7 +93,7 @@ MIDDLEWARE = [
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
-ROOT_URLCONF = "covsonar_backend.urls"
+ROOT_URLCONF = "sonar_backend.urls"
 CORS_ALLOWED_ORIGINS = [
     origin.strip()
     for origin in env("CORS_ALLOWED_ORIGINS", default="http://localhost:5173").split(
@@ -117,7 +117,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "covsonar_backend.wsgi.application"
+WSGI_APPLICATION = "sonar_backend.wsgi.application"
 
 
 # Database
