+++ apps/cli/tests/sars-cov-2/out.vcf
@@ -1,5 +1,5 @@
 ##fileformat=VCFv4.2
-##poweredby=covSonar2
+##poweredby=sonar
 ##reference=MN908947.3
 ##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype"
 #CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	IMS-10004-CVDP-0672526C-BAEA-4FE9-A57B-941CBCC13343	IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95	IMS-10013-CVDP-69DF29F4-D7E3-4954-94F4-65C20BE7B850	IMS-10013-CVDP-69DF29F4-D7E3-4954-94F4-65C20BE7B850,IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95	IMS-10025-CVDP-00960	IMS-10025-CVDP-00960,IMS-10004-CVDP-0672526C-BAEA-4FE9-A57B-941CBCC13343,IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95	IMS-10025-CVDP-00960,IMS-10013-CVDP-69DF29F4-D7E3-4954-94F4-65C20BE7B850,IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95	IMS-10025-CVDP-00960,IMS-10087-CVDP-D484F3AD-CD8F-473C-8A5E-DB5D6A710BE5,IMS-10004-CVDP-0672526C-BAEA-4FE9-A57B-941CBCC13343,IMS-10013-CVDP-69DF29F4-D7E3-4954-94F4-65C20BE7B850,IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95	IMS-10025-CVDP-00960,IMS-10087-CVDP-D484F3AD-CD8F-473C-8A5E-DB5D6A710BE5,IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95	IMS-10025-CVDP-00960,IMS-10087-CVDP-D484F3AD-CD8F-473C-8A5E-DB5D6A710BE5,IMS-10013-CVDP-69DF29F4-D7E3-4954-94F4-65C20BE7B850,IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95	IMS-10087-CVDP-D484F3AD-CD8F-473C-8A5E-DB5D6A710BE5	IMS-10087-CVDP-D484F3AD-CD8F-473C-8A5E-DB5D6A710BE5,IMS-10004-CVDP-0672526C-BAEA-4FE9-A57B-941CBCC13343,IMS-10013-CVDP-69DF29F4-D7E3-4954-94F4-65C20BE7B850,IMS-10013-CVDP-37E0BD5A-03D8-42CE-95C0-7B900B714B95
