+++ example-deploy/backend.env.example
@@ -1,7 +1,7 @@
 DEBUG=FALSE
 PROFILE_IMPORT=FALSE
 
-POSTGRES_DB=covsonar
+POSTGRES_DB=sonar
 POSTGRES_USER=postgres
 POSTGRES_HOST=sonar-db
 POSTGRES_PORT=5432
@@ -27,4 +27,4 @@ PROPERTY_BATCH_SIZE=1000
 
 GUNI_THREADS=2
 GUNI_WORKERS=2
-START_DJANGO_SERVER=python manage.py collectstatic --noinput && gunicorn --workers ${GUNI_WORKERS} --threads ${GUNI_THREADS} --bind 0.0.0.0:9080 covsonar_backend.wsgi:application
+START_DJANGO_SERVER=python manage.py collectstatic --noinput && gunicorn --workers ${GUNI_WORKERS} --threads ${GUNI_THREADS} --bind 0.0.0.0:9080 sonar_backend.wsgi:application
