+++ example-deploy/compose.yml
@@ -74,7 +74,7 @@ services:
     env_file:
       - ./backend.env
       - ./backend.secrets.env
-    command: celery -A covsonar_backend worker --loglevel ${LOG_LEVEL:-INFO} -Ofair --concurrency=4 -E --max-tasks-per-child 8 --time-limit 600 -n Bob
+    command: celery -A sonar_backend worker --loglevel ${LOG_LEVEL:-INFO} -Ofair --concurrency=4 -E --max-tasks-per-child 8 --time-limit 600 -n Bob
     volumes:
       - sonar-data:/sonar/data
       - sonar-logs:/logs
